+++ tests/integ/test_session.py
@@ -5,7 +5,6 @@
 
 import os
 from functools import partial
-from unittest.mock import Mock
 
 import pytest
 
@@ -316,9 +315,7 @@ def test_list_files_in_stage(session, resources_path):
 
 @pytest.mark.localtest
 @pytest.mark.skipif(IS_IN_STORED_PROC, reason="Cannot create session in SP")
-def test_create_session_from_parameters(
-    db_parameters, sql_simplifier_enabled, local_testing_mode
-):
+def test_create_session_from_parameters(db_parameters, sql_simplifier_enabled):
     session_builder = Session.builder.configs(db_parameters)
     new_session = session_builder.create()
     new_session.sql_simplifier_enabled = sql_simplifier_enabled
@@ -326,11 +323,8 @@ def test_create_session_from_parameters(
         df = new_session.createDataFrame([[1, 2]], schema=["a", "b"])
         Utils.check_answer(df, [Row(1, 2)])
         assert session_builder._options.get("password") is None
-        if not local_testing_mode:
-            assert new_session._conn._lower_case_parameters.get("password") is None
-            assert new_session._conn._conn._password is None
-        else:
-            assert isinstance(new_session._conn._conn._password, Mock)
+        assert new_session._conn._lower_case_parameters.get("password") is None
+        assert new_session._conn._conn._password is None
     finally:
         new_session.close()
 
@@ -534,10 +528,19 @@ def test_use_schema(db_parameters, sql_simplifier_enabled, local_testing_mode):
     del parameters["warehouse"]
     parameters["local_testing"] = local_testing_mode
     with Session.builder.configs(parameters).create() as session:
-        session.sql_simplifier_enabled = sql_simplifier_enabled
-        schema_name = db_parameters["schema"]
-        session.use_schema(schema_name)
-        assert session.get_current_schema() == f'"{schema_name.upper()}"'
+        quoted_schema_name = f'"SCHEMA_{Utils.random_alphanumeric_str(5)}_schema"'
+        try:
+            session.sql_simplifier_enabled = sql_simplifier_enabled
+            schema_name = db_parameters["schema"]
+            session.use_schema(schema_name)
+            assert session.get_current_schema() == f'"{schema_name.upper()}"'
+            if not local_testing_mode:
+                session.sql(f"CREATE OR REPLACE SCHEMA {quoted_schema_name}").collect()
+            session.use_schema(quoted_schema_name)
+            assert session.get_current_schema() == quoted_schema_name
+        finally:
+            if not local_testing_mode:
+                session.sql(f"DROP SCHEMA IF EXISTS {quoted_schema_name}").collect()
 
 
 @pytest.mark.localtest