+++ CHANGELOG.md
@@ -17,6 +17,10 @@
     - to_varchar
   - snowflake.snowpark.DataFrame:
     - pivot
+  - snowflake.snowpark.Session:
+    - cancel_all
+- Introduced a new exception class `snowflake.snowpark.mock.exceptions.SnowparkLocalTestingException`.
+- Added support for casting to FloatType
 
 #### Bug Fixes
 
@@ -30,6 +34,14 @@
 - Fixed a bug that on Windows platform that file operations was unable to properly handle file separator in directory name.
 - Fixed a bug that on Windows platform that when reading a pandas dataframe, IntervalType column with integer data can not be processed.
 - Fixed a bug that prevented users from being able to select multiple columns with the same alias.
+- Fixed a bug that `Session.get_current_[schema|database|role|user|account|warehouse]` returns upper-cased identifiers when identifiers are quoted.
+- Fixed a bug that function `substr` and `substring` can not handle 0-based `start_expr`.
+
+#### Improvements
+
+- Standardized the error experience by raising `SnowparkLocalTestingException` in error cases which is on par with `SnowparkSQLException` raised in non-local execution.
+- Improved error experience of `Session.write_pandas` method that `NotImplementError` will be raised when called.
+- Aligned error experience with reusing a closed session in non-local execution.
 
 ## 1.16.0 (TBD)